+++ importers/src/Service/VendorService/BogPortalen/BogPortalenVendorService.php
@@ -120,7 +120,7 @@ class BogPortalenVendorService implements VendorServiceImporterInterface, Vendor
     /**
      * {@inheritDoc}
      */
-    public function getUnverifiedVendorImageItem(string $identifier, string $type): ?UnverifiedVendorImageItem
+    public function getUnverifiedVendorImageItems(string $identifier, string $type): \Generator
     {
         if (!$this->supportsIdentifier($identifier, $type)) {
             throw new UnsupportedIdentifierTypeException(\sprintf('Unsupported single identifier: %s (%s)', $identifier, $type));
@@ -132,7 +132,7 @@ class BogPortalenVendorService implements VendorServiceImporterInterface, Vendor
         $item->setIdentifier($identifier);
         $item->setIdentifierType($type);
 
-        return $item;
+        yield $item;
     }
 
     /**