+++ importers/CHANGELOG.md
@@ -19,6 +19,7 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 - Simplify default validation to be based only on HTTP response code.
 - Allow for custom validators by vendor. This ensures non-standard behavior is scope to the relevant Vendor
 - Change `SingleIdentifier` vendors supports function to consider both type and identifier
+- Changed `VendorServiceSingleIdentifierInterface` to yield `UnverifiedVendorImageItem` items back  
 
 ### Added
 - Added ETag to check for "image updated"